// frontend/src/generated/models/OnboardingData.ts
export interface OnboardingData {
    /**
     * 
     * @type {string}
     * @memberof OnboardingData
     */
    sport: string;
    /**
     * 
     * @type {string}
     * @memberof OnboardingData
     */
    experienceLevel: string;
    /**
     * 
     * @type {number}
     * @memberof OnboardingData
     */
    weeklyTrainingHours: number;
    /**
     * 
     * @type {boolean}
     * @memberof OnboardingData
     */
    preparingForRace: boolean;
    /**
     * 
     * @type {Array<Race>}
     * @memberof OnboardingData
     */
    races: Array<Race>;
    /**
     * 
     * @type {string}
     * @memberof OnboardingData
     */
    currentVolume: string;
    /**
     * 
     * @type {string}
     * @memberof OnboardingData
     */
    longestRun: string;
    /**
     * 
     * @type {string}
     * @memberof OnboardingData
     */
    recentRaces: string;
    /**
     * 
     * @type {string}
     * @memberof OnboardingData
     */
    injuries: string;
    /**
     * 
     * @type {number}
     * @memberof OnboardingData
     */
    trainingDays: number;
    /**
     * 
     * @type {string}
     * @memberof OnboardingData
     */
    preferredWorkoutTime: string;
    /**
     * 
     * @type {boolean}
     * @memberof OnboardingData
     */
    gymAccess: boolean;
    /**
     * 
     * @type {Array<string>}
     * @memberof OnboardingData
     */
    crossTraining: Array<string>;
}